# sdk/java: per module cache