Append the module name to the maven cache volume.
We are seeing more and more troubles with the Java SDK and corruption of some files in the maven local repository (cached). This looks like to happen because of concurrent access to the maven repository... that are not supported.

Error looks like:

    [ERROR] Failed to execute goal
    org.apache.maven.plugins:maven-install-plugin:3.1.2:install
    (default-install) on project dagger-sdk-parent: Failed to install
    metadata io.dagger:dagger-sdk-parent/maven-metadata.xml:
    Could not parse metadata
    /root/.m2/repository/io/dagger/dagger-sdk-parent/maven-metadata-local.xml:
    in epilog non whitespace content is not allowed but got 0 (position:
    END_TAG seen ...</metadata>\n0... @13:2)

By adding the module name to the cache, this should limit the issue to happen while still keeping a cache for better performances.

Here are the time it takes to run a `dagger functions` 3 times in a row with and without cache:

Without cache: 28.5s | 11.3s | 0.7s
With cache: 11.1s | 6.6s | 0.7s
With cache with name: 10.8s | 6.7s | 0.7s

This should help to solve #10156