@helderco 2025-04-30
Most of our tests use the same name for modules. You could use the module source digest instead.