@helderco 2025-04-30
There's also options on cache volumes to make it private and not shared for example.