@eunomie 2025-04-30
> Most of our tests use the same name for modules. You could use the module source digest instead.
> There's also options on cache volumes to make it private and not shared for example.

🤔 yes, if I just use the name it will not solve all cases.

Digest might be wrong as if the module content change the cache will be invalid and that's not what we want.

I'll have a look at the `namespace` and to find a more precise way to specify the volume.