@eunomie 2025-04-30
I've updated it to use the module path/sub-path. That should cover most of the cases I guess.
That way the cache is only tight to a local directory, and not shared across multiple modules.

I wonder if I shouldn't put the path as the namespace, maybe that could be more elegant.