@helderco 2025-04-30
> I've updated it to use the module path/sub-path. That should cover most of the cases I guess. That way the cache is only tight to a local directory, and not shared across multiple modules.
> 
> I wonder if I shouldn't put the path as the namespace, maybe that could be more elegant.

It's like the name. Most of our tests create a module at `/work/test`. Same path and same name. Even concurrently.

Have you tried the sharing mode on the monted cache like I said? For example:

https://github.com/dagger/dagger/blob/37955449d1b32196d5be0f98c412c7bb5d7ff772/core/integration/provision_test.go#L194-L196