@eunomie 2025-05-02
> Have you tried the sharing mode on the monted cache like I said? For example:

Ho right, I only tried with the cache volume option, not the one on the mount cache function.

So yes, that looks better with lock mode. It will keep good performances when single access, and ensure we don't have corruption. It might slow down a bit when multiple concurrent access, but that should still be better than without cache so I think that's worth it.